format

+++ contracts/AvatarFactory.sol
@@ -46,7 +46,7 @@ contract ExposedAvatarFactory {
 
     mapping(uint256 => address) public avatarToOwner;
 
-    function createAvatarAndGetId() public returns (uint) {
+    function createAvatarAndGetId() public returns (uint256) {
         avatars.push(Avatar(0, 10, 10, 0));
         uint256 id = avatars.length - 1;
         avatarToOwner[id] = msg.sender;
